// src/types/types.d.ts
export interface Student{
    studentName: string;
    rollNumber: string;
    studentPhone: string;
    adminId: string;
    studentEmail: string;
    studentFatherName: string;
    studentAddress: string;
    studentYear: string;
    studentBranch: string;
    studentPassword: string;
    studentFatherPhone: string;
    feeStatus?:boolean;
}